requirements.txt
tlslite-ng
scapy
mitmproxy
pyshark
